Enable custom configuration
Also generalized migration of deprecated configuration properties
